cleaned up directory structure. Added config files and updated python folders tomore interpretable names
