+++ .gitignore
@@ -148,3 +148,6 @@ WindowsNoEditor
 RCode/ShinyApp/Data/*.csv
 PythonCode/PythonGridMapping/AirSimPythonClient/rav_*_mapper.py
 **/GPSMappings/Images/
+
+BlogRuntime/*
+JavaRuntime/*
