+++ BlogScripts/ChemicalProbablisticReasoning.blog
@@ -4,7 +4,7 @@ obs smell_camphor=false;
 obs smell_wood=false;
 obs smell_chlorine=false;
 obs smell_mustard_radish_garlic=false;
-obs smell_unpleasant=false;
+obs smell_unpleasant=true;
 obs smell_fish=false;
 obs smell_herring=false;
 obs smell_soap=false;
