+++ Config/config.ini
@@ -1,0 +1,32 @@
+[JAVA]
+JavaBinLoc = /JavaRuntime/jre1.8.0_171/bin/
+JavaCodeLoc = /JavaCode/
+JavaMissionDesignerJar = generateCoordinatesRevised.jar
+
+[BLOG]
+BlogBinLoc = /BlogRuntime/bin/
+BlogCodeLoc = /BlogScripts/
+BlogProgName = ChemicalProbablisticReasoning.blog
+
+[PYTHON]
+PythonDocRetrievalCodeLoc = /SOPRanking/RocsafeCode/Demo-IR/
+PythonRAVRouteExecutionDir = /PythonCode/Routing/AirSimPythonClient/RAVExecuteRoutes/
+PythonElasticMain = elasticMain.py
+PythonGenerateRouteDir = /PythonCode/Routing/RouteGeneration/
+PythonGenerateRoutesFileLoc = generateRoutesForUnreal.py
+
+[BATCHSCRIPTS]
+BatchScriptsLoc = /BatchScripts/
+
+[DATA]
+RAVGPSRoutesDir = /RCode/ShinyApp/Data/
+CollectedPNGImagesDir = /RAVCollectedData/PNGImages/
+RAVRecordedGPSWaypoints = /RAVCollectedData/GPSCoords/
+JavaCalculatedWaypointsLoc = /RCode/ShinyApp/Data/gridPoints.csv
+UIImagesDir = RCode/ShinyApp/Data/Images/
+MostRecentProcessImageFileLoc = mostRecentImageProcessed.jpg
+MostRecentImageRawFileLoc = mostRecentImageProcessed.jpg
+AgentRoutesDir = /RCode/ShinyApp/Data/
+
+
+
