+++ PythonCode/Routing/AirSimPythonClient/GPS/__init__.py
@@ -1,0 +1,2 @@
+from GPSCoordinate import GPSCoordinate
+__all__ = ['RAVGPSMapper', 'GPSCoordinate']
