--- PythonCode/Routing/AirSimPythonClient/GPSMappings/__init__.py
@@ -1,2 +1,0 @@
-from GPSCoordinate import GPSCoordinate
-__all__ = ['RAVGPSMapper', 'GPSCoordinate']
