+++ PythonCode/config.ini
@@ -1,8 +1,5 @@
 [DEFAULT]
-
 RAVGPSRoutesDir = "/RCode/ShinyApp/Data/"
 PNGImagesDir = "/RAVCollectedData/PNGImages/"
 RAVRecordedGPSWaypoints = "/RAVCollectedData/GPSCoords/"
-
-
 RAVRouteExecutionDir = "/PythonCode/Routing/AirSimPythonClient/RAVExecuteRoutes/"
