+++ RCode/ShinyApp/Data/Agent2.csv
@@ -1,37 +1,25 @@
 lat, long, alt
 53.28600000000000136424205265939235687255859375, -9.0587999999999997413624441833235323429107666015625, 20
-53.28049559433579946473400923423469066619873046875, -9.0620822568324498417761788005009293556213378906250, 100
-53.28049559433579946473400923423469066619873046875, -9.0623305527324498417761788005009293556213378906250, 100
-53.28049559433579946473400923423469066619873046875, -9.0625788486324498417761788005009293556213378906250, 100
-53.28049559433579946473400923423469066619873046875, -9.0628271445324498417761788005009293556213378906250, 100
-53.28049559433579946473400923423469066619873046875, -9.0630754404324498417761788005009293556213378906250, 100
-53.28049559433579946473400923423469066619873046875, -9.0633237363324498417761788005009293556213378906250, 100
-53.28029673433579946473400923423469066619873046875, -9.0633237363324498417761788005009293556213378906250, 100
-53.28029673433579946473400923423469066619873046875, -9.0635720322324498417761788005009293556213378906250, 100
-53.28029673433579946473400923423469066619873046875, -9.0638203281324498417761788005009293556213378906250, 100
-53.28029673433579946473400923423469066619873046875, -9.0640686240324498417761788005009293556213378906250, 100
-53.28029673433579946473400923423469066619873046875, -9.0643169199324498417761788005009293556213378906250, 100
-53.27989901433579946473400923423469066619873046875, -9.0643169199324498417761788005009293556213378906250, 100
-53.27989901433579946473400923423469066619873046875, -9.0645652158324498417761788005009293556213378906250, 100
-53.27989901433579946473400923423469066619873046875, -9.0648135117324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0645652158324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0643169199324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0640686240324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0638203281324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0635720322324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0633237363324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0630754404324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0628271445324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0625788486324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0623305527324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0620822568324498417761788005009293556213378906250, 100
-53.27970015433579946473400923423469066619873046875, -9.0618339609324498417761788005009293556213378906250, 100
-53.27989901433579946473400923423469066619873046875, -9.0618339609324498417761788005009293556213378906250, 100
-53.28029673433579946473400923423469066619873046875, -9.0620822568324498417761788005009293556213378906250, 100
-53.28029673433579946473400923423469066619873046875, -9.0623305527324498417761788005009293556213378906250, 100
-53.28029673433579946473400923423469066619873046875, -9.0625788486324498417761788005009293556213378906250, 100
-53.28029673433579946473400923423469066619873046875, -9.0628271445324498417761788005009293556213378906250, 100
-53.28029673433579946473400923423469066619873046875, -9.0630754404324498417761788005009293556213378906250, 100
-53.27950129433579946473400923423469066619873046875, -9.0630754404324498417761788005009293556213378906250, 100
-53.27950129433579946473400923423469066619873046875, -9.0625788486324498417761788005009293556213378906250, 100
+53.28044748450880133156970259733498096466064453125, -9.0624917635963902714636788005009293556213378906250, 100
+53.28025824660880133156970259733498096466064453125, -9.0624917635963902714636788005009293556213378906250, 100
+53.28006900870880133156970259733498096466064453125, -9.0624917635963902714636788005009293556213378906250, 100
+53.27987977080880133156970259733498096466064453125, -9.0624917635963902714636788005009293556213378906250, 100
+53.27969053290880133156970259733498096466064453125, -9.0624917635963902714636788005009293556213378906250, 100
+53.27950129500880133156970259733498096466064453125, -9.0624917635963902714636788005009293556213378906250, 100
+53.27931205710880133156970259733498096466064453125, -9.0624917635963902714636788005009293556213378906250, 100
+53.27931205710880133156970259733498096466064453125, -9.0628116779963902714636788005009293556213378906250, 100
+53.27950129500880133156970259733498096466064453125, -9.0628116779963902714636788005009293556213378906250, 100
+53.27969053290880133156970259733498096466064453125, -9.0628116779963902714636788005009293556213378906250, 100
+53.27987977080880133156970259733498096466064453125, -9.0628116779963902714636788005009293556213378906250, 100
+53.28006900870880133156970259733498096466064453125, -9.0628116779963902714636788005009293556213378906250, 100
+53.28025824660880133156970259733498096466064453125, -9.0628116779963902714636788005009293556213378906250, 100
+53.28044748450880133156970259733498096466064453125, -9.0628116779963902714636788005009293556213378906250, 100
+53.28044748450880133156970259733498096466064453125, -9.0631315923963902714636788005009293556213378906250, 100
+53.28025824660880133156970259733498096466064453125, -9.0631315923963902714636788005009293556213378906250, 100
+53.28006900870880133156970259733498096466064453125, -9.0631315923963902714636788005009293556213378906250, 100
+53.28006900870880133156970259733498096466064453125, -9.0634515067963902714636788005009293556213378906250, 100
+53.28025824660880133156970259733498096466064453125, -9.0634515067963902714636788005009293556213378906250, 100
+53.28025824660880133156970259733498096466064453125, -9.0621718491963902714636788005009293556213378906250, 100
+53.28006900870880133156970259733498096466064453125, -9.0621718491963902714636788005009293556213378906250, 100
+53.28006900870880133156970259733498096466064453125, -9.0618519347963902714636788005009293556213378906250, 100
